Encapsulate fetches into hook
